# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOC Triage Dashboard API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  open: 'open',
  analyzing: 'analyzing',
  analyzed: 'analyzed',
  closed: 'closed',
} as const;

export interface Session {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: SessionStatus;
  logCount: number;
  hasReport?: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface SessionInput {
  /** @minLength 1 */
  title: string;
  description?: string;
}

export type SessionUpdateStatus = typeof SessionUpdateStatus[keyof typeof SessionUpdateStatus];


export const SessionUpdateStatus = {
  open: 'open',
  analyzing: 'analyzing',
  analyzed: 'analyzed',
  closed: 'closed',
} as const;

export interface SessionUpdate {
  title?: string;
  description?: string;
  status?: SessionUpdateStatus;
}

export type SessionDetailStatus = typeof SessionDetailStatus[keyof typeof SessionDetailStatus];


export const SessionDetailStatus = {
  open: 'open',
  analyzing: 'analyzing',
  analyzed: 'analyzed',
  closed: 'closed',
} as const;

export type LogEntrySource = typeof LogEntrySource[keyof typeof LogEntrySource];


export const LogEntrySource = {
  fortigate: 'fortigate',
  watchguard: 'watchguard',
  agent_windows: 'agent_windows',
  agent_linux: 'agent_linux',
  unknown: 'unknown',
} as const;

export interface LogEntry {
  id: number;
  sessionId: number;
  source: LogEntrySource;
  rawJson: string;
  /** @nullable */
  extractedIp: string | null;
  /** @nullable */
  dstIp?: string | null;
  /** @nullable */
  dstPort?: number | null;
  /** @nullable */
  protocol?: string | null;
  /** @nullable */
  actionTaken?: string | null;
  /** @nullable */
  logTimestamp?: string | null;
  /** @nullable */
  ipType?: string | null;
  masked: boolean;
  createdAt: string;
}

export type ReportSeverity = typeof ReportSeverity[keyof typeof ReportSeverity];


export const ReportSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  informational: 'informational',
} as const;

/**
 * @nullable
 */
export type ReportVerdict = typeof ReportVerdict[keyof typeof ReportVerdict] | null;


export const ReportVerdict = {
  true_positive: 'true_positive',
  false_positive: 'false_positive',
  need_investigation: 'need_investigation',
} as const;

export interface Report {
  id: number;
  sessionId: number;
  summary: string;
  severity: ReportSeverity;
  /** @nullable */
  verdict?: ReportVerdict;
  /** @nullable */
  detailedFindings?: string | null;
  /** @nullable */
  technicalAnalysis?: string | null;
  /** @nullable */
  actionTaken?: string | null;
  iocs: string[];
  recommendations: string[];
  /** @nullable */
  attackVector?: string | null;
  affectedSystems?: string[];
  mitreAttackTechniques?: string[];
  rawAiResponse: string;
  /** @nullable */
  n8nExecutionId?: string | null;
  createdAt: string;
}

export interface SessionDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: SessionDetailStatus;
  logs: LogEntry[];
  report?: Report;
  createdAt: string;
  updatedAt: string;
}

export type LogEntryInputSource = typeof LogEntryInputSource[keyof typeof LogEntryInputSource];


export const LogEntryInputSource = {
  fortigate: 'fortigate',
  watchguard: 'watchguard',
  agent_windows: 'agent_windows',
  agent_linux: 'agent_linux',
  unknown: 'unknown',
} as const;

export interface LogEntryInput {
  source: LogEntryInputSource;
  rawJson: string;
}

export type IpCorrelationRiskLevel = typeof IpCorrelationRiskLevel[keyof typeof IpCorrelationRiskLevel];


export const IpCorrelationRiskLevel = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type IpCorrelationActionSummary = {
  blocked: number;
  allowed: number;
  detected: number;
  other: number;
};

export interface IpCorrelation {
  ip: string;
  maskedIp: string;
  ipType: string;
  sources: string[];
  logCount: number;
  /** 0-100 threat score computed from multi-source correlation, port risk, and action analysis */
  threatScore: number;
  riskLevel: IpCorrelationRiskLevel;
  portsSeen?: number[];
  actionSummary?: IpCorrelationActionSummary;
  logs: LogEntry[];
}

export interface AnalyzeInput {
  maskIps: boolean;
  additionalContext?: string;
}

export interface DashboardStats {
  totalSessions: number;
  openSessions: number;
  analyzedSessions: number;
  totalReports: number;
  totalLogs: number;
  criticalCount: number;
  highCount: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  session_created: 'session_created',
  session_analyzed: 'session_analyzed',
  report_generated: 'report_generated',
  log_added: 'log_added',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  /** @nullable */
  severity?: string | null;
  timestamp: string;
}

export interface ThreatBreakdownItem {
  severity: string;
  count: number;
}

export interface SourceDistributionItem {
  source: string;
  count: number;
}

export interface N8nConfig {
  /** @nullable */
  webhookUrl: string | null;
  isConfigured: boolean;
}

export interface N8nConfigInput {
  webhookUrl: string;
}