# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: SOC Triage Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Report:
      type: object
      required: [id, sessionId, summary, severity, iocs, recommendations, rawAiResponse, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        summary:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, informational]
        verdict:
          type: ["string", "null"]
          enum: [true_positive, false_positive, need_investigation, null]
        detailedFindings:
          type: ["string", "null"]
        technicalAnalysis:
          type: ["string", "null"]
        actionTaken:
          type: ["string", "null"]
        iocs:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        attackVector:
          type: ["string", "null"]
        affectedSystems:
          type: array
          items:
            type: string
        mitreAttackTechniques:
          type: array
          items:
            type: string
        rawAiResponse:
          type: string
        n8nExecutionId:
          type: ["string", "null"]
        createdAt:
          type: string
...