# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOC Triage Dashboard API
 * OpenAPI spec version: 0.2.0
 */
import type { ReportSeverity } from './reportSeverity';
import type { ReportVerdict } from './reportVerdict';

export interface Report {
  id: number;
  sessionId: number;
  summary: string;
  severity: ReportSeverity;
  /** @nullable */
  verdict?: ReportVerdict;
  /** @nullable */
  detailedFindings?: string | null;
  /** @nullable */
  technicalAnalysis?: string | null;
  /** @nullable */
  actionTaken?: string | null;
  iocs: string[];
  recommendations: string[];
  /** @nullable */
  attackVector?: string | null;
  affectedSystems?: string[];
  mitreAttackTechniques?: string[];
  rawAiResponse: string;
  /** @nullable */
  n8nExecutionId?: string | null;
  createdAt: string;
}